Refactor FXIOS-16432 [WebCompat] Strip the screenshot out of Technical Data
The thumbnail belongs to Report Preview, so Technical Data no longer carries
its section, view model fields, or layout branch. Nothing changes on screen:
the thumbnail stays off until FXIOS-16450 wires up the screenshot viewer.

Also drops the section cell's private copies of the card constants in favour
of the shared ones, and merges the two preview sample-page renderers.
