+++ BrowserKit/Sources/WebCompatReporterKit/WebCompatTechnicalDataViewController.swift
@@ -9,14 +9,10 @@ import UIKit
 @MainActor
 public protocol WebCompatTechnicalDataDelegate: AnyObject {
     func webCompatTechnicalDataDidRequestDismiss()
-    func webCompatTechnicalDataDidTapScreenshot()
 }
 
-/// The Technical Data screen: a tappable page thumbnail above collapsible sections listing the
-/// raw payload as key/value pairs. Store-agnostic, so configure it with a view model.
-///
-/// TODO: FXIOS-16432 - Add the plain-language Report Preview screen that pushes this one, and
-/// move the thumbnail onto it.
+/// The Technical Data screen, pushed from Report Preview: collapsible sections listing the raw
+/// payload as key/value pairs. Store-agnostic, so configure it with a view model.
 public final class WebCompatTechnicalDataViewController: UIViewController,
                                                          Themeable,
                                                          UICollectionViewDelegate {
@@ -28,13 +24,9 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
     /// `header` omits `rows` deliberately. Carrying them would mark the header changed on any
     /// value edit, when it only renders the title.
     private enum ItemKind: Equatable {
-        case screenshot(UIImage)
         case header(title: String, a11yIdentifier: String)
         case content(WebCompatTechnicalDataViewModel.PreviewSection)
     }
-
-    private static let screenshotDataSourceSectionID = "webcompat.preview.screenshot.section"
-    private static let screenshotDataSourceItemID = "webcompat.preview.screenshot.item"
 
     /// Suffixed so an item id can't collide with a section id in `itemsByID`.
     private static func headerDataSourceItemID(for sectionID: String) -> String {
@@ -53,17 +45,12 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
     private let notificationCenter: NotificationProtocol
 
     private var viewModel: WebCompatTechnicalDataViewModel
-    private var screenshot: UIImage?
     private var theme: Theme
 
     /// The data source keys on item ids, so the values themselves live here.
     private var itemsByID: [String: ItemKind] = [:]
-    /// Display order, so the layout closure can spot the screenshot section.
+    /// Display order, so an unchanged section set can skip the top-level apply.
     private var orderedSectionIDs: [String] = []
-
-    /// The first applyTheme runs inside `viewDidLoad`, where reconfiguring collapses the nav
-    /// bar's large title. Later changes do reconfigure.
-    private var hasAppliedThemeOnce = false
 
     private lazy var closeButton: UIBarButtonItem = {
         let image = UIImage(named: StandardImageIdentifiers.Large.cross)?.withRenderingMode(.alwaysTemplate)
@@ -102,9 +89,10 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
         super.viewDidLoad()
         setupNavigationItem()
         setupLayout()
-        configure(with: viewModel)
         listenForThemeChanges(withNotificationCenter: notificationCenter)
+        // Before the first configure, so the reconfigure hits an empty snapshot and no-ops.
         applyTheme()
+        configure(with: viewModel)
     }
 
     // MARK: - Configuration
@@ -116,14 +104,6 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
         navigationItem.title = viewModel.title
         closeButton.accessibilityLabel = viewModel.closeAccessibilityLabel
         closeButton.accessibilityIdentifier = viewModel.closeA11yIdentifier
-        updateSections()
-    }
-
-    /// The capture arrives after the screen is on screen, so it comes in on its own rather than
-    /// through the view model. Nil drops the thumbnail.
-    public func updateScreenshot(_ image: UIImage?) {
-        screenshot = image
-        guard isViewLoaded else { return }
         updateSections()
     }
 
@@ -140,17 +120,7 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
     }
 
     private func makeLayout() -> UICollectionViewCompositionalLayout {
-        return UICollectionViewCompositionalLayout { [weak self] index, environment in
-            let sectionIDs = self?.orderedSectionIDs ?? []
-            let sectionID = index < sectionIDs.count ? sectionIDs[index] : nil
-            if sectionID == Self.screenshotDataSourceSectionID {
-                var configuration = UICollectionLayoutListConfiguration(appearance: .plain)
-                configuration.backgroundColor = .clear
-                configuration.showsSeparators = false
-                let section = NSCollectionLayoutSection.list(using: configuration, layoutEnvironment: environment)
-                section.contentInsets.top = WebCompatReporterUX.Thumbnail.topInset
-                return section
-            }
+        return UICollectionViewCompositionalLayout { _, environment in
             var configuration = UICollectionLayoutListConfiguration(appearance: .insetGrouped)
             // Clear so the collection view's own themed background shows through. A theme change
             // then repaints instead of having to rebuild the layout for its colour.
@@ -164,20 +134,6 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
 
     // Registrations go here, not inside the provider: UIKit crashes on one built lazily per cell.
     private func makeDataSource() -> UICollectionViewDiffableDataSource<String, String> {
-        let screenshotRegistration = UICollectionView.CellRegistration<
-            WebCompatPreviewScreenshotCell, UIImage
-        > { [weak self] cell, _, image in
-            guard let self else { return }
-            cell.configure(
-                image: image,
-                imageAccessibilityLabel: self.viewModel.screenshotAccessibilityLabel,
-                a11yIdentifier: self.viewModel.screenshotA11yIdentifier
-            ) { [weak self] in
-                self?.delegate?.webCompatTechnicalDataDidTapScreenshot()
-            }
-            cell.applyTheme(theme: self.theme)
-        }
-
         let headerRegistration = UICollectionView.CellRegistration<
             UICollectionViewListCell, ItemKind
         > { [weak self] cell, _, item in
@@ -222,10 +178,6 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
                 )
             }
             switch self.itemsByID[itemID] {
-            case let .screenshot(image):
-                return collectionView.dequeueConfiguredReusableCell(
-                    using: screenshotRegistration, for: indexPath, item: image
-                )
             case let .header(title, a11yIdentifier):
                 return collectionView.dequeueConfiguredReusableCell(
                     using: headerRegistration,
@@ -252,10 +204,6 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
         itemsByID = [:]
         orderedSectionIDs = []
 
-        if let screenshot {
-            itemsByID[Self.screenshotDataSourceItemID] = .screenshot(screenshot)
-            orderedSectionIDs.append(Self.screenshotDataSourceSectionID)
-        }
         for section in viewModel.sections {
             itemsByID[Self.headerDataSourceItemID(for: section.id)] = .header(
                 title: section.title,
@@ -271,13 +219,6 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
             var sectionsSnapshot = NSDiffableDataSourceSnapshot<String, String>()
             sectionsSnapshot.appendSections(orderedSectionIDs)
             dataSource.apply(sectionsSnapshot, animatingDifferences: false)
-        }
-
-        if screenshot != nil,
-           dataSource.snapshot(for: Self.screenshotDataSourceSectionID).items.isEmpty {
-            var snapshot = NSDiffableDataSourceSectionSnapshot<String>()
-            snapshot.append([Self.screenshotDataSourceItemID])
-            dataSource.apply(snapshot, to: Self.screenshotDataSourceSectionID, animatingDifferences: false)
         }
 
         for section in viewModel.sections where dataSource.snapshot(for: section.id).items.isEmpty {
@@ -363,7 +304,6 @@ public final class WebCompatTechnicalDataViewController: UIViewController,
         view.backgroundColor = theme.colors.layer1
         navigationController?.navigationBar.tintColor = theme.colors.actionPrimary
         collectionView.backgroundColor = theme.colors.layer1
-        if hasAppliedThemeOnce { reconfigureAllItems() }
-        hasAppliedThemeOnce = true
+        reconfigureAllItems()
     }
 }
